+++ lib/zoi/types/time.ex
@@ -108,7 +108,7 @@ defmodule Zoi.Types.Time do
       case Time.compare(input, value) do
         :gt -> :ok
         :eq -> :ok
-        :lt -> {:error, Zoi.Error.greater_than_or_equal_to(:date, value, opts)}
+        :lt -> {:error, Zoi.Error.greater_than_or_equal_to(:time, value, opts)}
       end
     end
   end
@@ -122,7 +122,7 @@ defmodule Zoi.Types.Time do
       case Time.compare(input, value) do
         :lt -> :ok
         :eq -> :ok
-        :gt -> {:error, Zoi.Error.less_than_or_equal_to(:date, value, opts)}
+        :gt -> {:error, Zoi.Error.less_than_or_equal_to(:time, value, opts)}
       end
     end
   end
@@ -135,7 +135,7 @@ defmodule Zoi.Types.Time do
     def validate(_schema, input, value, opts) do
       case Time.compare(input, value) do
         :gt -> :ok
-        _ -> {:error, Zoi.Error.greater_than(:date, value, opts)}
+        _ -> {:error, Zoi.Error.greater_than(:time, value, opts)}
       end
     end
   end
@@ -148,7 +148,7 @@ defmodule Zoi.Types.Time do
     def validate(_schema, input, value, opts) do
       case Time.compare(input, value) do
         :lt -> :ok
-        _ -> {:error, Zoi.Error.less_than(:date, value, opts)}
+        _ -> {:error, Zoi.Error.less_than(:time, value, opts)}
       end
     end
   end